pyproject.toml
[project]
name = "mmae"
version = "0.1.0"
description = "Music/Media Audio Editor - Download, visualize, and clip audio from YouTube"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "yt-dlp>=2024.12.0",
    "pydub>=0.25.1",
    "python-multipart>=0.0.18",
    "audioop-lts>=0.2.1; python_version>='3.13'",
    "boto3>=1.35.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mmae = "main:run"
